requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
datetime>=5.2
json5>=0.9.14
urllib3>=2.0.7
certifi>=2023.11.17
charset-normalizer>=3.3.2
idna>=3.6
openai
google-generativeai # For Google Gemini
langchain
langchain-core
langchain-community
langgraph
graphviz
python-dotenv
langsmith
openai>=1.0.0
anthropic>=0.15.0
langchain_openai
langchain_google_genai
langchain_groq
fpdf>=2.0.0
reportlab
# Dataset utilities
# datasets>=2.0.0
# sentence-transformers
# pandas


